requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
jwt==1.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
sqlparse==0.5.3
typing_extensions==4.12.2
